Update concept space imports after moving types to space package
